+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Master
+- Feature: Added `snapshot` method to store to get raw value
+
 # 2.1.0-beta.3 4/3/18
 - Fix: Revert devtools trigger, due to circular